Redefine allowed option-chars
Allowed option chars are now all ASCII characters, except
control chars and whitespaces.
For single char options also the minus is forbidden and for
string option the minus can't be the first character.

Internally it would be possible to allow the full UTF-8 charset,
but a option '-😀' seems nonsense and also brings portabillity
problems.