Remove option -l/--local
Calling parseargs within a function and keeping the variables local to the
function can be achieved with 'typeset' or 'local' key words. No need to
handle this in parseargs.

With this also the special support for dash shell is removed, as the
difference between sh and dash is just the 'local' keyword.